+++ data/group-chat/src/main/java/com/seugi/data/groupchat/GroupChatRepository.kt
@@ -8,7 +8,7 @@ interface GroupChatRepository {
 
     suspend fun getGroupRoomList(workspaceId: String): Flow<Result<List<ChatRoomModel>>>
 
-    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Int>, chatRoomImg: String): Flow<Result<String>>
+    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Long>, chatRoomImg: String): Flow<Result<String>>
 
     suspend fun getGroupRoom(roomId: String): Flow<Result<ChatRoomModel>>
