+++ data/message/src/main/java/com/seugi/data/message/mapper/MessageLoadMapper.kt
@@ -4,7 +4,7 @@ import com.seugi.data.message.model.MessageLoadModel
 import com.seugi.network.message.response.message.MessageLoadResponse
 import kotlinx.collections.immutable.toImmutableList
 
-fun MessageLoadResponse.toModel(userId: Int) = MessageLoadModel(
+fun MessageLoadResponse.toModel(userId: Long) = MessageLoadModel(
     firstMessageId = firstMessageId,
     messages = messages.map { it.toModel(userId) }.toImmutableList(),
 )
