+++ data/perosnal-chat/src/main/java/com/seugi/data/personalchat/PersonalChatRepository.kt
@@ -8,7 +8,7 @@ interface PersonalChatRepository {
 
     suspend fun getAllChat(workspaceId: String): Flow<Result<List<ChatRoomModel>>>
 
-    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Int>, chatRoomImg: String): Flow<Result<String>>
+    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Long>, chatRoomImg: String): Flow<Result<String>>
 
     suspend fun getChat(roomId: String): Flow<Result<ChatRoomModel>>
 }
