+++ feature-main/chat-datail/src/main/java/com/seugi/chatdatail/navigation/ChatDetailNavigation.kt
@@ -20,7 +20,7 @@ fun NavController.navigateToChatDetail(
     navOptions,
 )
 
-fun NavGraphBuilder.chatDetailScreen(userId: Int, navigateToChatDetail: (workspaceId: String, chatRoomId: String) -> Unit, popBackStack: () -> Unit) {
+fun NavGraphBuilder.chatDetailScreen(userId: Long, navigateToChatDetail: (workspaceId: String, chatRoomId: String) -> Unit, popBackStack: () -> Unit) {
     composable(
         route = "$CHAT_DETAIL_ROUTE/{workspaceId}/{isPersonal}/{chatRoomId}",
         arguments = listOf(
