+++ feature-main/room-create/src/main/java/com/seugi/roomcreate/RoomCreateViewModel.kt
@@ -36,7 +36,7 @@ class RoomCreateViewModel @Inject constructor(
     private val _sideEffect = Channel<RoomCreateSideEffect>()
     val sideEffect = _sideEffect.receiveAsFlow()
 
-    fun loadUser(workspaceId: String, userId: Int) = viewModelScope.launch(dispatcher) {
+    fun loadUser(workspaceId: String, userId: Long) = viewModelScope.launch(dispatcher) {
         _state.update {
             it.copy(
                 isLoading = true,
@@ -69,7 +69,7 @@ class RoomCreateViewModel @Inject constructor(
         }
     }
 
-    fun updateChecked(userId: Int) {
+    fun updateChecked(userId: Long) {
         _state.value = _state.value.copy(
             userItem = _state.value.userItem.map {
                 if (it.id == userId) {
