+++ network/personal-chat/src/main/java/com/seugi/network/personalchat/PersonalChatDataSource.kt
@@ -7,7 +7,7 @@ interface PersonalChatDataSource {
 
     suspend fun getAllChat(workspaceId: String): BaseResponse<List<ChatRoomResponse>>
 
-    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Int>, chatRoomImg: String): BaseResponse<String>
+    suspend fun createChat(workspaceId: String, roomName: String, joinUsers: List<Long>, chatRoomImg: String): BaseResponse<String>
 
     suspend fun getChat(roomId: String): BaseResponse<ChatRoomResponse>
 }
